Refactor #17
