+++ src/QiqRenderer.php
@@ -19,6 +19,8 @@ use function substr;
 
 final class QiqRenderer implements RenderInterface
 {
+    private const LENGTH_OF_RESOURCE_DIR = 13;
+
     public function __construct(
         private TemplateCore $template
     ) {
@@ -26,24 +28,20 @@ final class QiqRenderer implements RenderInterface
 
     public function render(ResourceObject $ro): string
     {
-        $class = $this->getReflection($ro);
-        $fileName = $class->getFileName();
-        assert(is_string($fileName));
-        $path = $this->getTemplatePath($fileName);
-        $tpl = $this->template;
-        $tpl->setView($path);
+        $template = clone $this->template;
+        $this->setView($template, $ro);
         assert(is_array($ro->body));
-        $tpl->setData($ro->body);
+        $template->setData($ro->body);
 
-        return $tpl();
+        return ($template)();
     }
 
     /** @return ReflectionClass<object> */
     private function getReflection(ResourceObject $ro): ReflectionClass
     {
         if ($ro instanceof WeavedInterface) {
+            /** @var ReflectionClass<ResourceObject> $parentClass */
             $parentClass = (new ReflectionClass($ro))->getParentClass();
-            assert($parentClass instanceof ReflectionClass);
 
             return $parentClass;
         }
@@ -51,11 +49,15 @@ final class QiqRenderer implements RenderInterface
         return new ReflectionClass($ro);
     }
 
-    private function getTemplatePath(string $resourceFilePath): string
+    private function setView(TemplateCore $tpl, ResourceObject $ro): void
     {
-        $pos = strpos($resourceFilePath, '/Resource/');
-        $relativePath = substr($resourceFilePath, (int) $pos + 10);
+        $fileName = $this->getReflection($ro)->getFileName();
+        assert(is_string($fileName));
 
-        return str_replace('.php', '', $relativePath);
+        $pos = strpos($fileName, 'src/Resource/');
+        $relativePath = substr($fileName, (int) $pos + self::LENGTH_OF_RESOURCE_DIR);
+
+        $view = str_replace('.php', '', $relativePath);
+        $tpl->setView($view);
     }
 }
