+++ tests/Fake/FakeWeavedRo.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace BEAR\QiqModule;
 
-use BEAR\QiqModule\Fake\Resource\FakeRo;
+use BEAR\QiqModule\Resource\FakeRo;
 use Ray\Aop\WeavedInterface;
 
 class FakeWeavedRo extends FakeRo implements WeavedInterface
