+++ tests/Fake/templates/SubDirectory/FakeSub.php
